tidb/store/tikv: move withtikv to tikv/mockstore package
