+++ store/tikv/mockstore/test_util.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package util
+package mockstore
 
 import (
 	"flag"
