+++ store/tikv/tests/async_commit_fail_test.go
@@ -25,6 +25,7 @@ import (
 	"github.com/pingcap/parser/terror"
 	"github.com/pingcap/tidb/store/tikv"
 	tikverr "github.com/pingcap/tidb/store/tikv/error"
+	"github.com/pingcap/tidb/store/tikv/mockstore"
 	"github.com/pingcap/tidb/store/tikv/util"
 )
 
@@ -43,7 +44,7 @@ func (s *testAsyncCommitFailSuite) SetUpTest(c *C) {
 // committing primary region task.
 func (s *testAsyncCommitFailSuite) TestFailAsyncCommitPrewriteRpcErrors(c *C) {
 	// This test doesn't support tikv mode because it needs setting failpoint in unistore.
-	if *util.WithTiKV {
+	if *mockstore.WithTiKV {
 		return
 	}
 
@@ -75,7 +76,7 @@ func (s *testAsyncCommitFailSuite) TestFailAsyncCommitPrewriteRpcErrors(c *C) {
 
 func (s *testAsyncCommitFailSuite) TestAsyncCommitPrewriteCancelled(c *C) {
 	// This test doesn't support tikv mode because it needs setting failpoint in unistore.
-	if *util.WithTiKV {
+	if *mockstore.WithTiKV {
 		return
 	}
 
@@ -135,7 +136,7 @@ func (s *testAsyncCommitFailSuite) TestPointGetWithAsyncCommit(c *C) {
 
 func (s *testAsyncCommitFailSuite) TestSecondaryListInPrimaryLock(c *C) {
 	// This test doesn't support tikv mode.
-	if *util.WithTiKV {
+	if *mockstore.WithTiKV {
 		return
 	}
 
@@ -237,7 +238,7 @@ func (s *testAsyncCommitFailSuite) TestAsyncCommitContextCancelCausingUndetermin
 // TestAsyncCommitRPCErrorThenWriteConflict verifies that the determined failure error overwrites undetermined error.
 func (s *testAsyncCommitFailSuite) TestAsyncCommitRPCErrorThenWriteConflict(c *C) {
 	// This test doesn't support tikv mode because it needs setting failpoint in unistore.
-	if *util.WithTiKV {
+	if *mockstore.WithTiKV {
 		return
 	}
 
@@ -260,7 +261,7 @@ func (s *testAsyncCommitFailSuite) TestAsyncCommitRPCErrorThenWriteConflict(c *C
 // overwrites the undetermined error in the parent.
 func (s *testAsyncCommitFailSuite) TestAsyncCommitRPCErrorThenWriteConflictInChild(c *C) {
 	// This test doesn't support tikv mode because it needs setting failpoint in unistore.
-	if *util.WithTiKV {
+	if *mockstore.WithTiKV {
 		return
 	}
