+++ store/tikv/tests/async_commit_test.go
@@ -28,6 +28,7 @@ import (
 	"github.com/pingcap/tidb/store/mockstore/unistore"
 	"github.com/pingcap/tidb/store/tikv"
 	tikverr "github.com/pingcap/tidb/store/tikv/error"
+	"github.com/pingcap/tidb/store/tikv/mockstore"
 	"github.com/pingcap/tidb/store/tikv/mockstore/cluster"
 	"github.com/pingcap/tidb/store/tikv/oracle"
 	"github.com/pingcap/tidb/store/tikv/tikvrpc"
@@ -47,7 +48,7 @@ type testAsyncCommitCommon struct {
 }
 
 func (s *testAsyncCommitCommon) setUpTest(c *C) {
-	if *util.WithTiKV {
+	if *mockstore.WithTiKV {
 		s.store = NewTestStore(c)
 		return
 	}
@@ -195,7 +196,7 @@ func (s *testAsyncCommitSuite) lockKeysWithAsyncCommit(c *C, keys, values [][]by
 
 func (s *testAsyncCommitSuite) TestCheckSecondaries(c *C) {
 	// This test doesn't support tikv mode.
-	if *util.WithTiKV {
+	if *mockstore.WithTiKV {
 		return
 	}
 
@@ -402,7 +403,7 @@ func (s *testAsyncCommitSuite) TestAsyncCommitLinearizability(c *C) {
 // TestAsyncCommitWithMultiDC tests that async commit can only be enabled in global transactions
 func (s *testAsyncCommitSuite) TestAsyncCommitWithMultiDC(c *C) {
 	// It requires setting placement rules to run with TiKV
-	if *util.WithTiKV {
+	if *mockstore.WithTiKV {
 		return
 	}
