+++ store/tikv/tests/snapshot_fail_test.go
@@ -23,7 +23,7 @@ import (
 	"github.com/pingcap/tidb/store/mockstore/unistore"
 	"github.com/pingcap/tidb/store/tikv"
 	tikverr "github.com/pingcap/tidb/store/tikv/error"
-	"github.com/pingcap/tidb/store/tikv/util"
+	"github.com/pingcap/tidb/store/tikv/mockstore"
 )
 
 type testSnapshotFailSuite struct {
@@ -63,7 +63,7 @@ func (s *testSnapshotFailSuite) cleanup(c *C) {
 
 func (s *testSnapshotFailSuite) TestBatchGetResponseKeyError(c *C) {
 	// Meaningless to test with tikv because it has a mock key error
-	if *util.WithTiKV {
+	if *mockstore.WithTiKV {
 		return
 	}
 	defer s.cleanup(c)
@@ -92,7 +92,7 @@ func (s *testSnapshotFailSuite) TestBatchGetResponseKeyError(c *C) {
 
 func (s *testSnapshotFailSuite) TestScanResponseKeyError(c *C) {
 	// Meaningless to test with tikv because it has a mock key error
-	if *util.WithTiKV {
+	if *mockstore.WithTiKV {
 		return
 	}
 	defer s.cleanup(c)
